+++ index.js
@@ -15,8 +15,7 @@ const buildSchemas = require('./src/schemas');
 db.serialize(() => {
     buildSchemas(db);
     
-    const app = require('./src/app2');
+    const {app} = require('./src/app');
     require('./src/repository/RiderRepository').setDBInstance(db);
-
     app.listen(port, () => console.log(`App started and listening on port ${port}`));
 });
