+++ tests/logger.test.js
@@ -6,7 +6,7 @@ const request = require('supertest');
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database(':memory:');
 
-const app = require('../src/app')(db);
+// const app = require('../src/app')(db);
 const buildSchemas = require('../src/schemas');
 
 const {logger} = require('../src/services/winston');
